Add Nutrition Facts
A Package for NutritionFacts (to be associated with the product), NutritionFactsTable (all nutrition facts known by the Database) and NutritionFactsAdd (an option to insert a new tuple in the database) have been created. Every Package contains its own Fragment, View Model and ViewModelFactory.

The DAO has also been updated, a table for NutritionFacts has been added, as well as a table to associate nutrition facts with products.