Allow the user to add a custom Nutrition Fact
The addition of a new kind of Nutrition Fact is working (NutritionFactAdd). The new element is shown in nutritionFactTable, where the user can pick as many non-mandatory values to assign to the product. Those non mandatory nutrition facts picked are listed in NutritionFacts fragment, below the mandatories. The user can then associate a value to the amount contained for each nutrition fact. These values are saved when navigating to the next fragment.