Create Manufacturer DAO
Manufacturer has been added to the DB, as well as its DAO. ManufacturerViewModelFactory has been created to handle DB within the ViewModel. Once the user navigates from Manufacturer to the next fragment, the system will check if the name has already been registered. If not, it will create a new Manufacturer. In any case, it will associate the given Manufacturer with the current Product, updating idManufacturer in Product's table.